Add files via upload
DNAbp.txt contains the values and keys for the third function, which yield the complementary DNA sequence to the sequence put in by the user. The CTC... text file is an example file that is returned after running the function. The returned file contains the complementary sequence written inside of it, while the title of the file is the given/input sequence.